Refactor settings and other files
