+++ js/interface/settings.ts
@@ -1,7 +1,55 @@
-export const settings = {};
+import { Vue } from "../../lib/libs";
+import { Blockbench } from "../api";
+import { Dialog } from "./dialog";
+
+export const settings: Record<string, Setting> = {};
+
+type SettingsValue = string | number | boolean;
+interface SettingOptions {
+	name?: string
+	type?: 'number' | 'text' | 'toggle' | 'password' | 'select' | 'click'
+	value?: boolean | number | string
+	condition?: ConditionResolvable
+	category?: string
+	description?: string
+	requires_restart?: boolean
+	launch_setting?: boolean
+	//launch_setting?: boolean
+	min?: number
+	max?: number
+	step?: number
+	icon?: string
+	plugin?: string
+	click?(): void
+	options?: {
+		[id: string]: string
+	}
+	onChange?(value: any): void
+}
 
 export class Setting {
-	constructor(id, data) {
+	id: string
+	type: string
+	default_value: SettingsValue
+	master_value: SettingsValue
+	condition: ConditionResolvable
+	category: string
+	name: string
+	description: string
+	requires_restart: boolean
+	launch_setting: boolean
+	plugin?: string
+	min?: number
+	max?: number
+	step?: number
+	icon: string
+	click: (event: MouseEvent | KeyboardEvent) => void
+	options: Record<string, string>
+	hidden: boolean
+	onChange: (value: SettingsValue) => void
+	keybind_label: string
+
+	constructor(id: string, data: SettingOptions) {
 		this.id = id;
 		settings[id] = this;
 		this.type = 'toggle';
@@ -19,6 +67,7 @@ export class Setting {
 			}
 		}
 		if (typeof Settings.stored[id] === 'object') {
+			// @ts-ignore
 			this.master_value = Settings.stored[id].value;
 
 		} else if (data.value != undefined) {
@@ -91,7 +140,7 @@ export class Setting {
 		this.master_value = value;
 	}
 	get ui_value() {
-		let profile = Settings.dialog.content_vue?.profile;
+		let profile = Settings.dialog.content_vue?.$data.profile;
 		if (profile) {
 			return profile.settings[this.id] ?? this.master_value;
 		} else {
@@ -99,7 +148,7 @@ export class Setting {
 		}
 	}
 	set ui_value(value) {
-		let profile = Settings.dialog.content_vue?.profile;
+		let profile = Settings.dialog.content_vue?.$data.profile;
 		if (this.type == 'number') value = Math.clamp(value, this.min, this.max)
 		if (profile) {
 			Vue.set(profile.settings, this.id, value);
@@ -115,13 +164,13 @@ export class Setting {
 			delete Settings.structure[this.category].items[this.id];
 		}
 	}
-	set(value) {
+	set(value: SettingsValue) {
 		if (value === undefined || value === null) return;
 		let old_value = this.value;
 
 		if (this.type == 'number' && typeof value == 'number') {
-			if (this.snap) {
-				value = Math.round(value / this.snap) * this.snap;
+			if (this.step) {
+				value = Math.round(value / this.step) * this.step;
 			}
 			this.value = Math.clamp(value, this.min, this.max)
 		} else if (this.type == 'toggle') {
@@ -136,7 +185,7 @@ export class Setting {
 		}
 		Settings.saveLocalStorages();
 	}
-	trigger(e) {
+	trigger(e: KeyboardEvent | MouseEvent) {
 		let {type} = this;
 		let setting = this;
 		if (type == 'toggle') {
@@ -167,7 +216,7 @@ export class Setting {
 					}
 				})
 			}
-			new Menu(list).open(e.target);
+			new Menu(list).open(e.target as HTMLElement);
 
 		} else {
 			let dialog = new Dialog({
@@ -209,13 +258,32 @@ export class Setting {
 	}
 }
 
+enum SettingsProfileConditionType {
+	selectable = 'selectable',
+	format = 'format',
+	file_path = 'file_path',
+}
+interface SettingsProfileData {
+	name?: string
+	color?: number
+}
 export class SettingsProfile {
-	constructor(data = 0) {
+	uuid: string
+	name: string
+	color: number
+	condition: {
+		type: SettingsProfileConditionType,
+		value: string
+	}
+	settings: Record<string, SettingsValue>
+	selected: boolean
+
+	constructor(data: SettingsProfileData = {}) {
 		this.uuid = guid();
 		this.name = data.name || 'New Profile';
 		this.color = data.color == undefined ? Math.randomInteger(0, markerColors.length-1) : data.color;
 		this.condition = {
-			type: 'selectable',
+			type: SettingsProfileConditionType.selectable,
 			value: ''
 		};
 		this.settings = {};
@@ -224,7 +292,7 @@ export class SettingsProfile {
 		SettingsProfile.all.push(this);
 	}
 	select(update = true) {
-		if (this.condition.type !== 'selectable') return;
+		if (this.condition.type !== SettingsProfileConditionType.selectable) return;
 
 		SettingsProfile.all.forEach(p => p.selected = false);
 		this.selected = true;
@@ -252,12 +320,12 @@ export class SettingsProfile {
 	}
 	isActive() {
 		switch (this.condition.type) {
-			case 'selectable':
+			case SettingsProfileConditionType.selectable:
 				return SettingsProfile.selected == this;
-			case 'format':
+			case SettingsProfileConditionType.format:
 				if (Format && Format.id == this.condition.value) return true;
 				break;
-			case 'file_path':
+			case SettingsProfileConditionType.file_path:
 				let regex = new RegExp(this.condition.value, 'i');
 				if (Project && (
 					regex.test(Project.save_path.replace(osfs, '/')) ||
@@ -320,9 +388,9 @@ export class SettingsProfile {
 				remove: {type: 'buttons', buttons: ['generic.delete'], click: (button) => {
 					if (confirm(tl('settings_profile.confirm_delete')))
 					this.remove();
-					Settings.dialog.content_vue.profile = null;
+					Settings.dialog.content_vue.$data.profile = null;
 					SettingsProfile.unselect();
-					dialog.close();
+					dialog.close(0);
 				}}
 			},
 			onConfirm: (result) => {
@@ -343,331 +411,25 @@ export class SettingsProfile {
 		SettingsProfile.all.remove(this);
 		Settings.saveLocalStorages();
 	}
-}
-SettingsProfile.all = [];
-SettingsProfile.selected = null;
-SettingsProfile.unselect = function(update = true) {
-	SettingsProfile.all.forEach(p => p.selected = false);
-	SettingsProfile.selected = null;
-	if (update) {
-		Settings.updateSettingsInProfiles();
-		Settings.saveLocalStorages();
-		Settings.updateProfileButton();
+	static all: SettingsProfile[] = []
+	static selected: SettingsProfile | null = null
+	static unselect = function(update = true) {
+		SettingsProfile.all.forEach(p => p.selected = false);
+		SettingsProfile.selected = null;
+		if (update) {
+			Settings.updateSettingsInProfiles();
+			Settings.saveLocalStorages();
+			Settings.updateProfileButton();
+		}
 	}
 }
 
 export const Settings = {
+	profile_menu_button: null as HTMLElement | null,
 	structure: {},
-	stored: {},
-	setup() {
-		if (localStorage.getItem('settings') != null) {
-			Settings.stored = JSON.parse(localStorage.getItem('settings'));
-		}
-		
-		//General
-		new Setting('language', {value: 'en', type: 'select', requires_restart: true, options: Language.options});
-		new Setting('username', {value: '', type: 'text'});
-		new Setting('streamer_mode', {value: false, onChange() {
-			StartScreen.vue._data.redact_names = settings.streamer_mode.value;
-			Interface.status_bar.vue.streamer_mode = settings.streamer_mode.value;
-			updateStreamerModeNotification();
-		}});
-		new Setting('classroom_mode', {value: false, requires_restart: true});
-		new Setting('cdn_mirror', {value: false});
-		new Setting('recovery_save_interval', {value: 30, type: 'number', min: 0, onChange() {
-			clearTimeout(AutoBackup.loop_timeout);
-			AutoBackup.backupProjectLoop(false);
-		}});
-
-		//Interface
-		new Setting('interface_mode', 		{category: 'interface', value: 'auto', requires_restart: true, type: 'select', options: {
-			'auto': tl('settings.interface_mode.auto'),
-			'desktop': tl('settings.interface_mode.desktop'),
-			'mobile': tl('settings.interface_mode.mobile'),
-		}});
-		new Setting('interface_scale', 		{category: 'interface', value: 100, min: 40, max: 200, type: 'number', condition: isApp, onChange() {
-			var factor = Math.clamp(settings.interface_scale.value, 40, 200) / 100;
-			currentwindow.webContents.setZoomFactor(factor)
-			resizeWindow()
-		}});
-		new Setting('hide_tab_bar', 		{category: 'interface', value: Blockbench.isMobile, onChange() {
-			updateTabBarVisibility();
-		}});
-		new Setting('mobile_panel_side',	{category: 'interface', value: 'right', condition: Blockbench.isMobile, type: 'select', options: {
-			'right': tl('generic.right'),
-			'left': tl('generic.left'),
-		}, onChange() {
-			document.body.classList.toggle('mobile_sidebar_left', settings.mobile_panel_side.value == 'left');
-		}});
-		new Setting('status_bar_modifier_keys', {category: 'interface', value: true, condition: !Blockbench.isTouch, onChange(value) {
-			Interface.status_bar.vue.show_modifier_keys = value;
-		}});
-		new Setting('always_show_splash_art',{category: 'interface', value: true});
-		new Setting('origin_size',  		{category: 'interface', value: 10, type: 'number', min: 2, max: 40});
-		new Setting('control_size',  		{category: 'interface', value: 10, type: 'number', min: 2, max: 40});
-		new Setting('motion_trails',  		{category: 'interface', value: true, onChange() {
-			if (Animator.open) {
-				scene[this.value ? 'add' : 'remove'](Animator.motion_trail);
-			}
-		}});
-		new Setting('seethrough_outline', 	{category: 'interface', value: false, onChange(value) {
-			Canvas.outlineMaterial.depthTest = !value;
-			Canvas.meshOutlineMaterial.depthTest = !value;
-		}});
-		new Setting('outliner_colors', 		{category: 'interface', value: true});
-		new Setting('preview_checkerboard',	{category: 'interface', value: true, onChange() {
-			$('#center').toggleClass('checkerboard', settings.preview_checkerboard.value);
-		}});
-		new Setting('uv_checkerboard', 		{category: 'interface', value: true, onChange(val) {
-			UVEditor.vue.checkerboard = val;
-		}});
-		new Setting('timecode_frame_number',{category: 'interface', value: false, onChange() {
-			Timeline.vue.updateTimecodes();
-		}});
-		new Setting('only_selected_bezier_handles',{category: 'interface', value: false, onChange(val) {
-			Timeline.vue.show_all_handles = !val;
-		}});
-		
-		//Preview 
-		new Setting('brightness',  		{category: 'preview', value: 50, type: 'number', min: 0, max: 400, onChange() {
-			Canvas.updateShading();
-		}});
-		new Setting('shading', 	  		{category: 'preview', value: true, onChange() {
-			Canvas.updateShading()
-		}});
-		new Setting('antialiasing', 	{category: 'preview', value: true, requires_restart: true});
-		new Setting('antialiasing_bleed_fix', 	{category: 'preview', value: false, requires_restart: true});
-		new Setting('fov', 		  		{category: 'preview', value: 45, type: 'number', min: 1, max: 120, onChange(val) {
-			Preview.all.forEach(preview => preview.setFOV(val));
-		}});
-		new Setting('camera_near_plane',{category: 'preview', value: 1, type: 'number', min: 0.01, max: 100, onChange(val) {
-			Preview.all.forEach(preview => {
-				preview.camPers.near = val;
-				preview.camPers.updateProjectionMatrix();
-			});
-		}});
-		new Setting('render_sides', 			{category: 'preview', value: 'auto', type: 'select', options: {
-			'auto': tl('settings.render_sides.auto'),
-			'front': tl('settings.render_sides.front'),
-			'double': tl('settings.render_sides.double'),
-		}, onChange() {
-			Canvas.updateRenderSides();
-		}});
-		new Setting('tone_mapping', 			{category: 'preview', value: 'none', type: 'select', options: {
-			none: tl('generic.none'),
-			linear: 'Linear',
-			reinhard: 'Reinhard',
-			cineon: 'Cineon',
-			aces_filmic: 'ACES Filmic',
-			agx: 'AgX',
-			neutral: 'Neutral',
-		}, onChange() {
-			for (let preview of Preview.all) {
-				preview.updateToneMapping();
-			}
-			Canvas.updateShading();
-			for (let model of PreviewModel.getActiveModels()) {
-				model.material.needsUpdate = true;
-			}
-		}});
-		new Setting('fps_limit',				{category: 'preview', value: 144, min: 10, max: 1024, type: 'number'});
-		new Setting('background_rendering', 	{category: 'preview', value: true});
-		new Setting('texture_fps',   			{category: 'preview', value: 7, type: 'number', min: 0, max: 120, onChange() {
-			TextureAnimator.updateSpeed()
-		}});
-		new Setting('particle_tick_rate',		{category: 'preview', value: 30, type: 'number', min: 1, max: 1000, onChange() {
-			WinterskyScene.global_options.tick_rate = this.value;
-		}});
-		new Setting('volume', 					{category: 'preview', value: 80, min: 0, max: 200, type: 'number'});
-		new Setting('audio_scrubbing',			{category: 'preview', value: true});
-		new Setting('save_view_per_tab',		{category: 'preview', value: true});
-		new Setting('display_skin',				{category: 'preview', value: false, type: 'click', icon: 'icon-player', click: function() { changeDisplaySkin() }});
-
-		new Setting('viewport_rotate_speed',	{category: 'controls', value: 100, min: 10, max: 1000, type: 'number', onChange(value) {
-			Preview.all.forEach(viewport => viewport.controls.rotateSpeed = value / 100)
-		}});
-		new Setting('viewport_zoom_speed',		{category: 'controls', value: 100, min: 10, max: 1000, type: 'number', onChange(value) {
-			Preview.all.forEach(viewport => viewport.controls.zoomSpeed = value / 100 * 1.5)
-		}});
-		new Setting('editor_2d_zoom_speed',		{category: 'controls', value: 100, min: 10, max: 1000, type: 'number'});
-		new Setting('gamepad_controls',			{category: 'controls', value: false, name: 'Gamepad Controls', description: 'Use a gamepad or 3D mouse to navigate the viewport'});
-		new Setting('double_click_switch_tools',{category: 'controls', value: true});
-		new Setting('canvas_unselect',  		{category: 'controls', value: false});
-		new Setting('selection_tolerance', 		{category: 'controls', value: 10, type: 'number', min: 1, max: 50});
-
-		//Edit
-		new Setting('undo_selections',			{category: 'edit', value: false});
-		new Setting('undo_limit',				{category: 'edit', value: 256, type: 'number', min: 1});
-		new Setting('highlight_cubes',  		{category: 'edit', value: true, onChange() {
-			updateCubeHighlights();
-		}});
-		new Setting('outliner_reveal_on_select', {category: 'edit', value: true})
-		new Setting('allow_display_slot_mirror', {category: 'edit', value: false, onChange(value) {
-			DisplayMode.vue.allow_mirroring = value;
-		}})
-		new Setting('deactivate_size_limit',	{category: 'edit', value: false});
-		new Setting('modded_entity_integer_size',{category:'edit', value: true});
-		new Setting('vertex_merge_distance',	{category: 'edit', value: 0.1, step: 0.01, type: 'number', min: 0});
-		new Setting('preview_paste_behavior',	{category: 'edit', value: 'always_ask', type: 'select', options: {
-			'always_ask': tl('settings.preview_paste_behavior.always_ask'),
-			'outliner': tl('menu.paste.outliner'),
-			'face': tl('menu.paste.face'),
-			'mesh_selection': tl('menu.paste.mesh_selection'),
-		}});
-		new Setting('stretch_linked',		{category: 'edit', value: true});
-		new Setting('auto_keyframe',		{category: 'edit', value: true});
-
-		//Paint
-		new Setting('color_wheel',					{category: 'paint', value: false, onChange(value) {
-			Interface.Panels.color.vue.picker_type = value ? 'wheel' : 'box';
-		}});
-		new Setting('brush_cursor_2d',			{category: 'paint', value: true});
-		new Setting('brush_cursor_3d',			{category: 'paint', value: true, onChange(value) {
-			if (!value) scene.remove(Canvas.brush_outline);
-		}});
-		new Setting('outlines_in_paint_mode',		{category: 'paint', value: true});
-		new Setting('move_with_selection_tool',		{category: 'paint', value: true});
-		new Setting('pick_color_opacity',			{category: 'paint', value: false});
-		new Setting('paint_through_transparency',	{category: 'paint', value: true});
-		new Setting('paint_side_restrict',			{category: 'paint', value: true});
-		new Setting('paint_with_stylus_only',		{category: 'paint', value: false});
-		new Setting('brush_opacity_modifier',		{category: 'paint', value: 'none', type: 'select', options: {
-			'pressure': tl('settings.brush_modifier.pressure'),
-			'tilt': tl('settings.brush_modifier.tilt'),
-			'none': tl('settings.brush_modifier.none'),
-		}});
-		new Setting('brush_size_modifier', {category: 'paint', value: 'none', type: 'select', options: {
-			'pressure': tl('settings.brush_modifier.pressure'),
-			'tilt': tl('settings.brush_modifier.tilt'),
-			'none': tl('settings.brush_modifier.none'),
-		}});
-		new Setting('image_editor',  	{category: 'paint', value: false, type: 'click', condition: isApp, icon: 'fas.fa-pen-square', click: function() {changeImageEditor(null) }});
-		
-		//Grid
-		new Setting('grids',				{category: 'grid', value: true, onChange() {Canvas.buildGrid()}});
-		new Setting('base_grid',			{category: 'grid', value: true, onChange() {Canvas.buildGrid()}});
-		new Setting('large_grid', 			{category: 'grid', value: true, onChange() {Canvas.buildGrid()}});
-		new Setting('full_grid',			{category: 'grid', value: false, onChange() {Canvas.buildGrid()}});
-		new Setting('large_box',			{category: 'grid', value: false, onChange() {Canvas.buildGrid()}});
-		new Setting('large_grid_size',		{category: 'grid', value: 3, type: 'number', min: 0, max: 2000, onChange() {Canvas.buildGrid()}});
-		//new Setting('display_grid',		{category: 'grid', value: false});
-		new Setting('pixel_grid',			{category: 'grid', value: false, onChange(value) {
-			Canvas.updatePixelGrid();
-			UVEditor.vue.pixel_grid = value;
-		}});
-		new Setting('painting_grid',		{category: 'grid', value: true, onChange(value) {
-			Canvas.updatePixelGrid();
-			UVEditor.vue.pixel_grid = value;
-		}});
-		new Setting('image_editor_grid_size',{category: 'grid', type: 'number', value: 16, onChange() {
-			UVEditor.vue.zoom += 0.01;
-			UVEditor.vue.zoom -= 0.01;
-		}});
-		new Setting('ground_plane',			{category: 'grid', value: false, onChange() {
-			Canvas.ground_plane.visible = this.value;
-		}});
-		new Setting('ground_plane_double_side',{category: 'grid', value: false, onChange() {
-			Canvas.groundPlaneMaterial.side = this.value ? THREE.DoubleSide : THREE.FrontSide;
-		}});
-		
-		//Snapping
-		new Setting('edit_size',		{category: 'snapping', value: 16, type: 'number', min: 1, max: 8192, onChange() {Canvas.buildGrid()}});
-		new Setting('shift_size', 		{category: 'snapping', value: 64, type: 'number', min: 1, max: 8192});
-		new Setting('ctrl_size',		{category: 'snapping', value: 160, type: 'number', min: 1, max: 8192});
-		new Setting('ctrl_shift_size',	{category: 'snapping', value: 640, type: 'number', min: 1, max: 8192});
-		new Setting('negative_size',	{category: 'snapping', value: false});
-		new Setting('nearest_rectangle_select',{category: 'snapping', value: false});
-		
-		//Defaults
-		new Setting('default_cube_size',		{category: 'defaults', value: 2, type: 'number', min: 0, max: 32});
-		new Setting('autouv',					{category: 'defaults', value: true});
-		new Setting('inherit_parent_color',		{category: 'defaults', value: false});
-		new Setting('create_rename', 			{category: 'defaults', value: false});
-		new Setting('show_only_selected_uv', 	{category: 'defaults', value: false});
-		new Setting('default_path', 			{category: 'defaults', value: false, type: 'click', condition: isApp, icon: 'burst_mode', click: function() { openDefaultTexturePath() }});
-		new Setting('animation_snap',			{category: 'defaults', value: 24, type: 'number'});
-		new Setting('uniform_keyframe',			{category: 'defaults', value: true});
-		
-		//Dialogs
-		new Setting('dialog_larger_cubes', 		{category: 'dialogs', value: true, name: tl('message.model_clipping.title'), description: tl('settings.dialog.desc', [tl('message.model_clipping.title')])});
-		new Setting('dialog_rotation_limit', 	{category: 'dialogs', value: true, name: tl('message.rotation_limit.title'), description: tl('settings.dialog.desc', [tl('message.rotation_limit.title')])});
-		new Setting('dialog_loose_texture', 	{category: 'dialogs', value: true, name: tl('message.loose_texture.title'), description: tl('settings.dialog.desc', [tl('message.loose_texture.title')])});
-		new Setting('dialog_invalid_characters',{category: 'dialogs', value: true, name: tl('message.invalid_characters.title'), description: tl('settings.dialog.desc', [tl('message.invalid_characters.title')])});
-		new Setting('dialog_save_codec',		{category: 'dialogs', value: true, name: tl('message.save_codec_selector.title'), description: tl('settings.dialog.desc', [tl('message.save_codec_selector.title')])});
-		
-		//Application
-		new Setting('recent_projects', {category: 'application', value: 32, max: 256, min: 0, type: 'number', condition: isApp});
-		new Setting('backup_interval', {category: 'application', value: 10, type: 'number', min: 0, condition: isApp});
-		new Setting('backup_retain', {category: 'application', value: 30, type: 'number', min: 0, condition: isApp});
-		new Setting('automatic_updates', {category: 'application', value: true, condition: isApp});
-		new Setting('update_to_prereleases', {category: 'application', value: false, condition: isApp, launch_setting: true});
-		new Setting('hardware_acceleration', {category: 'application', value: true, requires_restart: true, condition: isApp, launch_setting: true});
-		
-		//Export
-		new Setting('json_indentation',		{category: 'export', value: 'tabs', type: 'select', options: {
-			tabs: tl('settings.json_indentation.tabs'),
-			spaces_4: tl('settings.json_indentation.spaces_4'),
-			spaces_2: tl('settings.json_indentation.spaces_2'),
-		}});
-		new Setting('final_newline',		{category: 'export', value: false});
-		new Setting('minifiedout',			{category: 'export', value: false});
-		new Setting('embed_textures', 		{category: 'export', value: true});
-		new Setting('minify_bbmodel', 		{category: 'export', value: true});
-		new Setting('export_empty_groups',	{category: 'export', value: true});
-		new Setting('export_groups', 		{category: 'export', value: true});
-		new Setting('java_export_pivots', 	{category: 'export', value: true});
-		new Setting('optifine_save_default_texture',{category: 'export', value: true});
-		new Setting('obj_face_export_mode',	{category: 'export', value: 'both', type: 'select', options: {
-			both: tl('settings.obj_face_export_mode.both'),
-			tris: tl('settings.obj_face_export_mode.tris'),
-			quads: tl('settings.obj_face_export_mode.quads'),
-		}});
-		new Setting('animation_sample_rate',{category: 'export', value: 24, type: 'number', min: 1, max: 640});
-		new Setting('model_export_scale',	{category: 'export', value: 16, type: 'number', min: 0.0001, max: 4096});
-		new Setting('sketchfab_token', 		{category: 'export', value: '', type: 'password'});
-		new Setting('credit', 				{category: 'export', value: 'Made with Blockbench', type: 'text'});
-
-		Blockbench.onUpdateTo('4.12.1', () => {
-			settings.antialiasing_bleed_fix.set(false);
-		})
-	},
-	setupProfiles() {
-		if (localStorage.getItem('settings_profiles') != null) {
-			let profiles = JSON.parse(localStorage.getItem('settings_profiles'));
-			profiles.forEach(profile => {
-				let new_profile = new SettingsProfile(profile);
-				if (profile.selected) new_profile.select(false);
-			})
-		}
-		Settings.profile_menu_button = document.getElementById('settings_profiles_header_menu');
-		Settings.profile_menu_button.addEventListener('click', event => {
-			let list = [
-				{
-					name: 'generic.none',
-					icon: SettingsProfile.selected ? 'far.fa-circle' : 'far.fa-dot-circle',
-					click: () => {
-						SettingsProfile.unselect();
-					}
-				},
-				'_'
-			];
-			SettingsProfile.all.forEach(profile => {
-				if (profile.condition.type != 'selectable') return;
-				list.push({
-					name: profile.name,
-					icon: profile.selected ? 'far.fa-dot-circle' : 'far.fa-circle',
-					color: markerColors[profile.color].standard,
-					click: () => {
-						profile.select();
-					}
-				})
-			})
-			new Menu(list).open(Settings.profile_menu_button);
-		});
-		Settings.profile_menu_button.setAttribute('title', tl('data.settings_profile'))
-		Settings.updateProfileButton();
-	},
-	addCategory(id, data = {}) {
+	stored: {} as Record<string, SettingsValue>,
+	dialog: null as Dialog | null,
+	addCategory(id: string, data: {name?: string, open?: boolean}) {
 		Settings.structure[id] = {
 			name: data.name || tl('settings.category.'+id),
 			open: data.open != undefined ? !!data.open : id === 'general',
@@ -684,24 +446,23 @@ export const Settings = {
 		localStorage.setItem('settings', JSON.stringify(settings_copy) )
 		localStorage.setItem('settings_profiles', JSON.stringify(SettingsProfile.all));
 
-		if (window.ColorPanel) {
-			ColorPanel.saveLocalStorages()
-		}
+		// @ts-ignore
+		if (window.ColorPanel) ColorPanel.saveLocalStorages()
 	},
 	save() {
 		Settings.saveLocalStorages()
 		updateSelection();
 		for (let key in BarItems) {
 			let action = BarItems[key]
-			if (action.linked_setting) {
+			if (action instanceof Toggle && action.linked_setting) {
 				if (settings[action.linked_setting] && action.value != settings[action.linked_setting].value) {
-					action.value = settings[action.linked_setting].value;
+					action.value = settings[action.linked_setting].value as boolean;
 					action.updateEnabledState();
 				}
 			}
 		}
 		Settings.updateProfileButton();
-		Blockbench.dispatchEvent('update_settings');
+		Blockbench.dispatchEvent('update_settings', {});
 	},
 	updateSettingsInProfiles() {
 		let settings_to_change = new Set();
@@ -714,7 +475,7 @@ export const Settings = {
 				}
 			}
 		}
-		settings_to_change.forEach(key => {
+		settings_to_change.forEach((key: string) => {
 			let setting = settings[key];
 			if (setting.onChange) setting.onChange(setting.value);
 		})
@@ -733,28 +494,28 @@ export const Settings = {
 			}
 		}
 	},
-	get(id) {
+	get(id: string) {
 		if (id && settings[id]) {
 			return settings[id].value;
 		}
 	},
-	openDialog(options = {}) {
+	openDialog(options: {search_term?: string, profile?: string} = {}) {
 		for (var sett in settings) {
 			if (settings.hasOwnProperty(sett)) {
 				Settings.old[sett] = settings[sett].value
 			}
 		}
 		Settings.dialog.show();
-		if (options.search_term) Settings.dialog.content_vue.search_term = options.search_term;
-		if (options.profile) Settings.dialog.content_vue.profile = options.profile;
+		if (options.search_term) Settings.dialog.content_vue.$data.search_term = options.search_term;
+		if (options.profile) Settings.dialog.content_vue.$data.profile = options.profile;
 		Settings.dialog.content_vue.$forceUpdate();
 	},
-	showRestartMessage(settings) {
+	showRestartMessage(settings?: Setting[]) {
 		let message;
 		if (settings instanceof Array) {
 			message = tl('message.settings_require_restart.message') + '\n\n';
-			for (let plugin of settings) {
-				message += '* ' + plugin.name + '\n'
+			for (let setting of settings) {
+				message += '* ' + setting.name + '\n'
 			}
 		}
 		Blockbench.showMessageBox({
@@ -771,7 +532,7 @@ export const Settings = {
 					Blockbench.once('before_closing', () => {
 						ipcRenderer.send('new-window');
 					})
-					currentwindow.close();
+					window.close();
 				} else {
 					location.reload();
 				}
@@ -779,24 +540,6 @@ export const Settings = {
 		})
 	},
 	old: {}
-}
-Settings.setup()
-
-export function updateStreamerModeNotification() {
-	$('#start_screen section#streamer_mode').detach()
-
-	if (settings.streamer_mode.value) {
-		addStartScreenSection('streamer_mode', {
-			graphic: {type: 'icon', icon: 'live_tv'},
-			color: 'var(--color-stream)',
-			text_color: 'var(--color-light)',
-			text: [
-				{type: 'h2', text: tl('interface.streamer_mode_on'), click() {
-					Settings.openDialog({search_term: tl('settings.streamer_mode')})
-				}}
-			]
-		})
-	}
 }
 
 BARS.defineActions(() => {
@@ -809,8 +552,8 @@ BARS.defineActions(() => {
 					Settings.old[sett] = settings[sett].value
 				}
 			}
-			Settings.dialog.show()
-			document.querySelector('dialog#settings .search_bar > input').focus()
+			Settings.dialog.show();
+			(document.querySelector('dialog#settings .search_bar > input') as HTMLElement).focus();
 		}
 	})
 	
@@ -818,6 +561,7 @@ BARS.defineActions(() => {
 		icon: 'folder',
 		category: 'blockbench',
 		click: function () {
+			// @ts-ignore for now
 			Blockbench.import({
 				resource_id: 'config',
 				extensions: ['bbsettings'],
@@ -856,6 +600,7 @@ BARS.defineActions(() => {
 				});
 				if (!go_on) return;
 			}
+			// @ts-ignore for now
 			Blockbench.export({
 				resource_id: 'config',
 				type: 'Blockbench Settings',
@@ -864,8 +609,9 @@ BARS.defineActions(() => {
 			})
 		}
 	})
-	BarItems.import_settings.toElement('#settings_title_bar')
-	BarItems.export_settings.toElement('#settings_title_bar')
+	let title_bar = document.getElementById('settings_title_bar');
+	BarItems.import_settings.toElement(title_bar);
+	BarItems.export_settings.toElement(title_bar);
 })
 
 onVueSetup(function() {
@@ -889,6 +635,13 @@ onVueSetup(function() {
 		sidebar_pages[key] = Settings.structure[key].name;
 	}
 
+	interface SettingsDialogVueData {
+		structure: any,
+		profile: null | SettingsProfile,
+		all_profiles: SettingsProfile[],
+		open_category: string,
+		search_term: string
+	}
 	Settings.dialog = new Dialog({
 		id: 'settings',
 		title: 'dialog.settings.settings',
@@ -919,12 +672,12 @@ onVueSetup(function() {
 				all_profiles: SettingsProfile.all,
 				open_category: 'general',
 				search_term: ''
-			}},
+			} as SettingsDialogVueData},
 			methods: {
-				saveSettings() {
+				saveSettings(this: SettingsDialogVueData) {
 					Settings.saveLocalStorages();
 				},
-				settingContextMenu(setting, event) {
+				settingContextMenu(setting: Setting, event: MouseEvent) {
 					new Menu([
 						{
 							name: 'dialog.settings.reset_to_default',
@@ -936,11 +689,12 @@ onVueSetup(function() {
 						}
 					]).open(event);
 				},
-				showProfileMenu(event) {
-					let items = [
+				showProfileMenu(this: SettingsDialogVueData) {
+					let items: MenuItem[] = [
 						{
 							name: 'generic.none',
 							icon: 'remove',
+							color: '',
 							click: () => {
 								this.profile = null;
 							}
@@ -969,15 +723,16 @@ onVueSetup(function() {
 							this.profile.openDialog();
 						}}
 					)
+					// @ts-ignore
 					new Menu('settings_profiles', items).open(this.$refs.profile_menu)
 				},
-				profileButtonPress() {
+				profileButtonPress(this: SettingsDialogVueData) {
 					if (!this.profile) {
 						this.profile = new SettingsProfile({});
 					}
 					this.profile.openDialog();
 				},
-				getProfileValuesForSetting(key) {
+				getProfileValuesForSetting(this: SettingsDialogVueData, key) {
 					return this.all_profiles.filter(profile => {
 						return profile.settings[key] !== undefined;
 					});
