+++ js/types.d.ts
@@ -1,4 +1,4 @@
-/// <reference types="blockbench-types" />
+/// <reference types="../../blockbench-types" />
 declare module "*.glsl" {
 	const value: string;
 	export default value;
