+++ js/util/global.d.ts
@@ -128,6 +128,8 @@ declare global {
 	interface Event {
 		shiftKey?: boolean
 	}
+
+	let osfs: string
 }
 
 export { }
